Seperated Client Socket Initialization
Created a function called, clientSocketInitialization. This function creates a socket that we can use to fire off packets to nodes. It reutrns a struct that holds the socket as well as info that is needed to for all UDP messages so we're not re-evaluating it. 

Send state now takes this struct along with the IP addresses to send the state to.