Protobuf parsing on server side is ~functional
On the client side for the message size I was passing the size of the pointer not the object....oops!

So now the server is receiving everything except the last two values. Better than the first three!

Removed the "required" flag from nodestate. It was causing debug headaches. Protobuf documentation recommended against using required flags for this reason.