Protobuf Server Prasing Complete
So in the node simulator I was passing the rand() function to a bool value. Wasnt thinking much of it.

What was happening was the rand() funciton was putting massive numbers into the bool values, which was bloating the size of the message. Due to this unexpected bloat protobuf was chopping off the end of the message, thus cutting off the state and temp fields.